% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.matrix.R
\name{adjacency.matrix}
\alias{adjacency.matrix}
\title{Adjacency matrix from amino acid or nucleotide sequences}
\usage{
adjacency.matrix(
  input.sequences = NULL,
  normalize = TRUE,
  sequence.dictionary = amino.acids[1:20]
)
}
\arguments{
\item{input.sequences}{The amino acid or nucleotide sequences to use}

\item{normalize}{Return the values as a function of total number of 
residues (TRUE) or frequencies (FALSE)}

\item{sequence.dictionary}{The letters to use in sequence generation 
(default are all amino acids)}
}
\value{
Adjacency matrix based on input.sequences.
}
\description{
Use this calculate frequency of adjacency between residues
along a set of biological sequences.
}
\examples{
new.sequences <- generate.sequences(prefix.motif = "CAS",
                                    suffix.motif = "YF",
                                    number.of.sequences = 100,
                                    min.length = 8,
                                    max.length = 16)
                          
sequence.matrix <- adjacency.matrix(new.sequences,
                                    normalize = TRUE)
                        
}
