% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{amino.acids}
\alias{amino.acids}
\title{Standard 20 amino acids}
\format{
An object of class \code{character} of length 20.
}
\usage{
amino.acids
}
\description{
Vector of one-letter codes for the 20 standard amino acids.
}
\keyword{datasets}
