% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apex_blosum.pam.matrices}
\alias{apex_blosum.pam.matrices}
\title{List of amino acid substitution matrices}
\description{
A list of amino acid substitution matrices, including:

* BLOSUM45
* BLOSUM50  
* BLOSUM62  
* BLOSUM80  
* BLOSUM100
* PAM30
* PAM40
* PAM70    
* PAM120
* PAM250
}
