% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apex_example.data}
\alias{apex_example.data}
\title{Example contig data for Apex}
\description{
A list of example contigs from the following sources:

* TenX: 10k_Human_DTC_Melanoma_5p_nextgem_Multiplex from 10x Website
* AIRR: Human_colon_16S8157851 from \href{https://pubmed.ncbi.nlm.nih.gov/37055623/}{PMID: 37055623}
* Adaptive: Adaptive_2283_D0 from \href{https://pubmed.ncbi.nlm.nih.gov/36220826/}{PMID: 36220826}
* Omniscope: Internal Data
}
\concept{Data}
