% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apex_gene.list}
\alias{apex_gene.list}
\title{A list of IMGT gene names by genes, loci, and species}
\description{
A list of gene nomenclature to use for converting for 
uniformity. Data is organize by gene region, loci and species.
 Data pulled from IMGT.org 09/05/2024.
}
