% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.sequences.R
\name{generate.sequences}
\alias{generate.sequences}
\title{Randomly Generate Amino Acid Sequences}
\usage{
generate.sequences(
  prefix.motif = NULL,
  suffix.motif = NULL,
  number.of.sequences = 100,
  min.length = 1,
  max.length = 10,
  sequence.dictionary = amino.acids[1:20]
)
}
\arguments{
\item{prefix.motif}{Add defined amino acid/nucleotide sequence to the start of the generated sequences.}

\item{suffix.motif}{Add defined amino acid/nucleotide sequence to the end of the generated sequences}

\item{number.of.sequences}{Number of sequences to generate}

\item{min.length}{Minimum length of the final sequence. The min.length may be adjusted if 
incongruent with prefix.motif/suffix.motif lengths}

\item{max.length}{Maximum length of the final sequence}

\item{sequence.dictionary}{The letters to use in sequence generation (default are all amino acids)}
}
\value{
A vector of generated sequences
}
\description{
Use this to make synthetic amino acid sequences
for purposes of testing code, training models,
or noise.
}
\examples{
generate.sequences(prefix.motif = "CAS",
                   suffix.motif = "YF",
                   number.of.sequences = 100,
                   min.length = 8,
                   max.length = 16)
                        
}
