% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.IMGT.R
\name{get.IMGT}
\alias{get.IMGT}
\title{Get IMGT Sequences for Specific Loci}
\usage{
get.IMGT(
  species = "human",
  chain = "TRB",
  sequence.type = "aa",
  frame = "inframe",
  region = "v"
)
}
\arguments{
\item{species}{One or two word designation of species. Currently supporting: 
"human", "mouse", "rat", "rabbit", "rhesus monkey", "sheep", "pig", "platypus",
"alpaca", "dog", "chicken", and "ferret"}

\item{chain}{Sequence chain to access}

\item{sequence.type}{Type of sequence - "aa" for amino acid or "nt" for nucleotide}

\item{frame}{Designation for "all", "inframe" or "inframe+gap"}

\item{region}{Sequence gene loci to access}
}
\value{
A list of allele sequences
}
\description{
Use this to access the ImMunoGeneTics (IMGT) sequences for a
specific species and gene loci. More information on 
IMGT can be found at \href{https://www.imgt.org/}{imgt.org}.
}
\examples{
TRBV_aa <- get.IMGT(species = "human",
                   chain = "TRB",
                   frame = "inframe",
                   region = "v",
                   sequence.type = "aa") 
                              
}
