% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{get_substitution_matrix}
\alias{get_substitution_matrix}
\title{Get substitution matrix from package data or custom input}
\usage{
get_substitution_matrix(matrix_name)
}
\arguments{
\item{matrix_name}{Character string or numeric matrix}
}
\value{
Numeric matrix with amino acid row/column names
}
\description{
Get substitution matrix from package data or custom input
}
\keyword{internal}
