% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIR.R
\name{getIR}
\alias{getIR}
\title{Extract Immune Receptor Sequences}
\usage{
getIR(input.data, chains, sequence.type = "aa")
}
\arguments{
\item{input.data}{Single-cell object or the output of 
\link[scRepertoire]{combineTCR} and \link[scRepertoire]{combineBCR} from
scRepertoire}

\item{chains}{Immune Receptor chain to use - \strong{TRA}, 
\strong{TRB}, \strong{IGH}, or \strong{IGL}}

\item{sequence.type}{Extract amino acid (\strong{aa}) or 
nucleotide (\strong{nt}) sequences}
}
\value{
A data frame of nucleotide or amino acid sequences
}
\description{
Use this to extract immune receptor sequences from a Single-Cell 
Object or the output of \link[scRepertoire]{combineTCR} and 
\link[scRepertoire]{combineBCR}.
}
