% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immApex-package.R
\docType{package}
\name{immApex-package}
\alias{immApex}
\alias{immApex-package}
\title{immApex: Tools for Adaptive Immune Receptor Sequence-Based Machine and Deep Learning}
\description{
A set of tools to for machine and deep learning in R from amino acid and nucleotide sequences focusing on adaptive immune receptors. The package includes pre-processing of sequences, unifying gene nomenclature usage, encoding sequences, and combining models. This package will serve as the basis of future immune receptor sequence functions/packages/models compatible with the scRepertoire ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BorchLab/immApex/}
  \item Report bugs at \url{https://github.com/BorchLab/immApex/issues}
}

}
\author{
\strong{Maintainer}: Nick Borcherding \email{ncborch@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Nick Borcherding \email{ncborch@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Qile Yang \email{qile.yang@berkeley.edu} (\href{https://orcid.org/0009-0005-0148-2499}{ORCID}) [contributor]
}

}
\keyword{internal}
