% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immapex_AA.data}
\alias{immapex_AA.data}
\title{A list of amino acid properties}
\description{
A list of amino acid properties that are
used for \code{\link{propertyEncoder}} function.

This includes:
\itemize{
  \item{atchleyFactors} 
  \item{crucianiProperties} 
  \item{FASGAI} 
  \item{kideraFactors} 
  \item{MSWHIM}
  \item{ProtFP} 
  \item{stScales}
  \item{tScales}
  \item{VHSE} 
  \item{zScales}
  }
}
