% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.cdr.R
\name{infer.cdr}
\alias{infer.cdr}
\title{Infer portions of the CDR Loop from Vgene data}
\usage{
infer.cdr(
  input.data,
  reference = NULL,
  chain = "TRB",
  technology = NULL,
  sequence.type = "aa",
  sequences = c("CDR1", "CDR2")
)
}
\arguments{
\item{input.data}{Data frame of sequencing data or 
output from format.genes().}

\item{reference}{IMGT sequences from get.IMGT()}

\item{technology}{The sequencing technology employed - 'TenX', "Adaptive', 
'AIRR', or 'Omniscope'}

\item{sequence.type}{Type of sequence - "aa" for amino acid or "nt" for nucleotide}

\item{sequences}{The specific regions of the CDR loop to get from the data.}
}
\value{
A data frame with the new columns of CDR sequences added.
}
\description{
Use this isolate sequences from the CDR loop using
the V gene annotation. When there are multiple 
V gene matches for a single gene, the first allelic 
sequence is used.
}
\examples{
# Getting the Sequence Reference
TRBV_aa <- get.IMGT(species = "human",
                    chain = "TRB",
                    frame = "inframe",
                    region = "v",
                    sequence.type = "aa") 
                    
# Ensuring sequences are formatted to IMGT                   
TenX_formatted <- format.genes(Apex_example.data[["TenX"]],
                               region = "v",
                               technology = "TenX")
             
# Inferring CDR loop elements           
TenX_formatted <- infer.cdr(TenX_formatted,
                            chain = "TRB", 
                            reference = TRBV_aa,
                            technology = "TenX", 
                            sequence.type = "aa",
                            sequences = c("CDR1", "CDR2")
                            
}
