% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{make_identity_matrix}
\alias{make_identity_matrix}
\title{Create a simple identity substitution matrix}
\usage{
make_identity_matrix()
}
\description{
Create a simple identity substitution matrix
}
\keyword{internal}
