% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.sequences.R
\name{mutate.sequences}
\alias{mutate.sequences}
\title{Randomly Mutate Sequences of Amino Acids}
\usage{
mutate.sequences(
  input.sequences,
  n.sequences = 1,
  mutation.rate = 0.01,
  position.start = NULL,
  position.end = NULL,
  sequence.dictionary = amino.acids[1:20]
)
}
\arguments{
\item{input.sequences}{The amino acid or nucleotide sequences to use}

\item{n.sequences}{The number of mutated sequences to return}

\item{mutation.rate}{The rate of mutations to introduce into sequences}

\item{position.start}{The starting position to mutate along the sequence. 
Default NULL will start the random mutations at position 1.}

\item{position.end}{The ending position to mutate along the sequence.
Default NULL will end the random mutations at the last position.}

\item{sequence.dictionary}{The letters to use in sequence mutation
(default are all amino acids)}
}
\value{
A vector of mutated sequences
}
\description{
Use this to mutate amino acid sequences
for purposes of testing code, training models,
or noise.
}
\examples{
sequences <- generate.sequences(prefix.motif = "CAS",
                                suffix.motif = "YF",
                                number.of.sequences = 100,
                                min.length = 8,
                                max.length = 16)
                                
mutated_sequences <- mutate.sequences(sequences, 
                                      n.sequence = 1,
                                      position.start = 3,
                                      position.end = 8)


}
