% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.hot.encoder.R
\name{one.hot.encoder}
\alias{one.hot.encoder}
\title{One Hot Encoder from Amino Acid String}
\usage{
one.hot.encoder(
  input.sequences,
  max.length = NULL,
  split.length = 1,
  convert.to.matrix = TRUE,
  sequence.dictionary = amino.acids[1:20]
)
}
\arguments{
\item{input.sequences}{The amino acid or nucleotide sequences to use}

\item{max.length}{Additional length to pad, NULL will pad sequences 
to the max length of input.sequences}

\item{convert.to.matrix}{Return a matrix (TRUE) or a 3D array (FALSE)}
}
\value{
One hot encoded amino acid sequences in a matrix or 3D array
}
\description{
Use this to transform amino acid sequences a
one hot encoding of the sequence.
}
\examples{
new.sequences <- generate.sequences(prefix.motif = "CAS",
                                    suffix.motif = "YF",
                                    number.of.sequences = 100,
                                    min.length = 8,
                                    max.length = 16)
                          
sequence.matrix <- one.hot.encoder(new.sequences, 
                                  convert.to.matrix = TRUE)
                        
}
