% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalEncoder.R
\name{positionalEncoder}
\alias{positionalEncoder}
\title{Adding Position-Specific Information to Sequences}
\usage{
positionalEncoder(number.of.sequences, latent.dims = NULL)
}
\arguments{
\item{number.of.sequences}{The number of sequences to generate 
position information}

\item{latent.dims}{The number of latent dimensions.}
}
\value{
A matrix of values
}
\description{
Use this calculate positional encoding for recurrent
neural networks using sin/cos and position information.
}
\examples{
                         
position.info <- positionalEncoder(number.of.sequences = 1000, 
                                   latent.dims = 64)
                        
}
