% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.sequences.R
\name{tokenize.sequences}
\alias{tokenize.sequences}
\title{Generate Tokenized Sequences from Amino Acid String}
\usage{
tokenize.sequences(
  input.sequences,
  add.startstop = TRUE,
  start.token = "!",
  stop.token = "^",
  max.length = NULL,
  convert.to.matrix = TRUE
)
}
\arguments{
\item{input.sequences}{The amino acid or nucleotide sequences to use}

\item{add.startstop}{Add start and stop tokens to the sequence}

\item{start.token}{The character to use for the start token}

\item{stop.token}{The character to use for the stop token}

\item{max.length}{Additional length to pad, NULL will pad sequences 
to the max length of input.sequences}

\item{convert.to.matrix}{Return a matrix (TRUE) or a vector (FALSE)}
}
\value{
Tokenized amino acid sequences in a matrix or vector
}
\description{
Use this to transform amino acid sequences into 
tokens in preparing for deep learning models.
}
\examples{
new.sequences <- generate.sequences(prefix.motif = "CAS",
                                    suffix.motif = "YF",
                                    number.of.sequences = 100,
                                    min.length = 8,
                                    max.length = 16)
                          
sequence.matrix <- tokenize.sequences(new.sequences, 
                                     add.startstop = TRUE,
                                     start.token = "!",
                                     stop.token = "^", 
                                     convert.to.matrix = TRUE)
                        
}
