% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apex_gene.list}
\alias{apex_gene.list}
\title{A list of IMGT gene names by genes, loci, and species}
\description{
A list of gene names to use for converting for 
uniformity. Data pulled from IMGT.org 09/05/2024.

A list of example contigs from the following sources:

* TenX: 10k_Human_DTC_Melanoma_5p_nextgem_Multiplex from 10x Website
* AIRR: Human_colon_16S8157851 from \href{https://pubmed.ncbi.nlm.nih.gov/37055623/}{PMID: 37055623}
* Adaptive: Adaptive_2283_D0 from \href{https://pubmed.ncbi.nlm.nih.gov/36220826/}{PMID: 36220826}
* Omniscope: Internal Data
}
\concept{Data}
