% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.encoder.R
\name{geometric.encoder}
\alias{geometric.encoder}
\title{Geometric Encoder from Amino Acid Strings}
\usage{
geometric.encoder(input.sequences, method.to.use = "BLOSUM62", theta = pi)
}
\arguments{
\item{input.sequences}{The amino acid sequences to use}

\item{method.to.use}{The method or approach to use for the conversion: 
"BLOSUM45", "BLOSUM50", "BLOSUM62", "BLOSUM80", "BLOSUM100", "PAM30", 
"PAM40", "PAM70", "PAM120", or "PAM250"}

\item{theta}{angle to use for geometric transformation}
}
\value{
Geometric encoded amino acid sequences in a matrix
}
\description{
Use this to transform amino acid sequences into a geometric 
encoding of the sequence.
}
\examples{
new.sequences <- generate.sequences(prefix.motif = "CAS",
                                    suffix.motif = "YF",
                                    number.of.sequences = 100,
                                    min.length = 8,
                                    max.length = 16)
                          
sequence.matrix <- one.hot.encoder(new.sequences, 
                                   convert.to.matrix = TRUE)
                        
}
