% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricEncoder.R
\name{geometricEncoder}
\alias{geometricEncoder}
\title{Geometric Encoder from Amino Acid Strings}
\usage{
geometricEncoder(input.sequences, method.to.use = "BLOSUM62", theta = pi/3)
}
\arguments{
\item{input.sequences}{The set of amino acid sequences}

\item{method.to.use}{The method or approach for the conversion: 
\itemize{
  \item{BLOcks SUbstitution Matrices: BLOSUM45, BLOSUM50, BLOSUM62, BLOSUM80, BLOSUM100}
  \item{Point Accepted Mutation Matrices: PAM30, PAM40, PAM70, PAM120, PAM250}
}}

\item{theta}{The angle for geometric transformation}
}
\value{
Geometric encoded amino acid sequences in a matrix
}
\description{
Use this to transform amino acid sequences into a geometric 
encoding of the sequence.
}
\examples{
new.sequences <- generateSequences(prefix.motif = "CAS",
                                   suffix.motif = "YF",
                                   number.of.sequences = 100,
                                   min.length = 8,
                                   max.length = 16)
                          
sequence.matrix <- geometricEncoder(new.sequences, 
                                    method.to.use = "BLOSUM62",
                                    theta = pi/3)
                        
}
