% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIMGT.R
\name{getIMGT}
\alias{getIMGT}
\title{Get IMGT Sequences for Specific Loci}
\usage{
getIMGT(
  species = "human",
  chain = "TRB",
  sequence.type = "aa",
  frame = "inframe",
  region = "v",
  verbose = TRUE
)
}
\arguments{
\item{species}{One or two word common designation of species. Currently supporting: 
\strong{human}, \strong{mouse}, \strong{rat}, \strong{rabbit}, \strong{rhesus monkey}, 
\strong{sheep}, \strong{pig}, \strong{platypus}, \strong{alpaca}, \strong{dog}, \strong{chicken}, 
and \strong{ferret}}

\item{chain}{Sequence chain to access, like \strong{TRB} or \strong{IGH}}

\item{sequence.type}{Type of sequence - \strong{aa} for amino acid or \strong{nt} for nucleotide}

\item{frame}{Designation for \strong{all}, \strong{inframe} or \strong{inframe+gap} that includes
aligned sequences with the IMGT gaps}

\item{region}{Sequence gene loci to access}

\item{verbose}{Print messages corresponding to the processing step}
}
\value{
A list of allele sequences
}
\description{
Use this to access the ImMunoGeneTics (IMGT) sequences for a
specific species and gene loci. More information on 
IMGT can be found at \href{https://www.imgt.org/}{imgt.org}.
}
\examples{
TRBV_aa <- getIMGT(species = "human",
                   chain = "TRB",
                   frame = "inframe",
                   region = "v",
                   sequence.type = "aa") 
                              
}
