% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immapex_blosum.pam.matrices}
\alias{immapex_blosum.pam.matrices}
\title{List of amino acid substitution matrices}
\description{
A list of amino acid substitution matrices, using the Point 
Accepted Matrix (PAM) and BLOck SUbstitution Matrix (BLOSUM) approaches. 
A discussion and comparison of these matrices are available at
\href{https://pubmed.ncbi.nlm.nih.gov/21356840/}{PMID: 21356840}.

\itemize{
 \item{BLOSUM45}
 \item{BLOSUM50}  
 \item{BLOSUM62}
 \item{BLOSUM80}
 \item{BLOSUM100}
 \item{PAM30}
 \item{PAM40}
 \item{PAM70}    
 \item{PAM120}
 \item{PAM250}
 }
}
