% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variationalSequences.R
\name{variationalSequences}
\alias{variationalSequences}
\title{Generate Similar Sequences using Variational Autoencoder}
\usage{
variationalSequences(
  input.sequences,
  number.of.sequences = 100,
  encoder = "onehotEncoder",
  aa.method.to.use = NULL,
  layers = 2,
  hidden.dims = c(256, 128),
  latent.dim = 16,
  batch.size = 16,
  epochs = 30,
  learning.rate = 0.001,
  epsilon.std = 1,
  null.threshold = 0.05,
  call.theshold = 0.5,
  activation.function = "relu",
  optimizer = "adam",
  sequence.dictionary = amino.acids[1:20]
)
}
\arguments{
\item{input.sequences}{The amino acid or nucleotide sequences to use}

\item{number.of.sequences}{Number of sequences to generate}

\item{encoder}{The method to prepare the sequencing information - 
"onehotEncoder" or "propertyEncoder"}

\item{aa.method.to.use}{The method or approach to use for the conversion:
\itemize{
  \item{Individual sets: atchleyFactors, crucianiProperties, FASGAI, kideraFactors, MSWHIM,
  ProtFP, stScales, tScales, VHSE, zScales"}
  \item{Multiple Sets: c("atchleyFactors", "VHSE") }
}}

\item{layers}{The number of hidden layers to employ within the VAE}

\item{hidden.dims}{A vector of the neurons to use in the hidden layers, 
The length needs to match the number of layers}

\item{latent.dim}{The size of the latent dimensions}

\item{batch.size}{The batch size to use for VAE training}

\item{epochs}{The number of epochs to use in VAE training}

\item{learning.rate}{The learning rate to use in VAE training}

\item{epsilon.std}{The epsilon to use in VAE training}

\item{null.threshold}{The null threshold to use in VAE training}

\item{activation.function}{The activation for the dense connected layers}

\item{optimizer}{The optimizer to use in VAE training}

\item{sequence.dictionary}{The letters to use in sequence mutation
(default are all amino acids)}
}
\value{
A vector of mutated sequences
}
\description{
Use this to simulate sequences using a variational 
autoencoder (VAE) and perturbation of the probability distributions.
}
\examples{
sequences <- generateSequences(prefix.motif = "CAS",
                               suffix.motif = "YF",
                               number.of.sequences = 100,
                               min.length = 8,
                               max.length = 16)
                                
variate_sequences <- variationalSequences(sequences, 
                                          encoder = "onehotEncoder",
                                          layers = 2, 
                                          hidden.dims = c(256, 128)
                                          latent.dim = 16,
                                          batch.size = 16)

}
